fix: bug on spending logic all was needed was to update the sql database to persist data helped by 2 devs

+++ Development Space/RegisterPortal/src/main/java/alvarico/allynn/User.java
@@ -69,7 +69,7 @@ public class User {
         return password;
     }
 
-    public float getBalance() {
+    public int getBalance() {
         return balance;
     }
 }
